requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# OpenAI - Funciona en Mac ARM64
openai>=1.3.0

# Web scraping y data fetching
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Presentation generation
python-pptx>=0.6.21
pillow>=10.0.0

# NLP y análisis de texto
nltk>=3.8.1
scikit-learn>=1.3.0
spacy>=3.6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Data formats
pyarrow>=12.0.0